% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality_4test.R
\docType{data}
\name{airquality_4test}
\alias{airquality_4test}
\title{Updated airquality for testing.}
\format{
A data frame with 100 observations and 8 columns:
\describe{
\item{ Ozone }{Ozone Score}
\item{ Solar.R }{Solar Score}
\item{ Wind }{Wind Score}
\item{ Temp }{Temperature}
\item{ Month }{Month of Year}
\item{ Day }{Day of Month}
\item{ CharVar1 }{Dummy Character Variable 1}
\item{ CharVar2 }{Dummy Character Variable 2}
}
}
\source{
Generated from repfun package.
}
\usage{
airquality_4test
}
\description{
airquality_4test
}
\details{
Updated version of the airquality data frame that includes filled NA values for testing.
}
\examples{
data("airquality_4test")
}
\references{
None
}
\keyword{datasets}
