% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airquality_updated.R
\docType{data}
\name{airquality_updated}
\alias{airquality_updated}
\title{Updated airquality}
\format{
A data frame with 100 observations and 8 columns:
\describe{
\item{ Ozone }{Ozone Score}
\item{ Solar.R }{Solar Score}
\item{ Wind }{Wind Score}
\item{ Temp }{Temperature}
\item{ Month }{Month of Year}
\item{ Day }{Day of Month}
\item{ CharVar1 }{Dummy Character Variable 1}
\item{ CharVar2 }{Dummy Character Variable 2}
}
}
\source{
Generated from repfun package.
}
\usage{
airquality_updated
}
\description{
airquality_updated
}
\details{
Updated version of the airquality data frame that includes NA values.
}
\examples{
data("airquality_updated")
}
\references{
None
}
\keyword{datasets}
