% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/copydata.R
\name{copydata}
\alias{copydata}
\title{Copy package data to the directory specified.}
\usage{
copydata(p)
}
\arguments{
\item{p}{A path as a string.}
}
\value{
'NULL' because package dataframes are copied to a folder.
}
\description{
Copy all package data into a temporary directory that can be used when running examples.
}
\examples{
library(repfun)
copydata(tempdir())

}
\author{
Yongwei Wang, \email{yongwei.x.wang@viivhealthcare.com} \cr
Chris Rook, \email{cr883296@gmail.com}
}
