% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/formats.R
\docType{data}
\name{formats}
\alias{formats}
\title{Data frame of SAS formats}
\format{
A data frame with 5 columns and 8 observations:
\describe{
\item{ FMTNAME }{Name of Format}
\item{ START }{Start Value of Format}
\item{ END }{End Value of Format}
\item{ LABEL }{Label of Format}
\item{ TYPE }{Type of Format}
}
}
\source{
Generated for repfun package.
}
\usage{
formats
}
\description{
formats
}
\details{
Formats Data Frame (SAS Style)
}
\examples{
data("formats")
}
\references{
None
}
\keyword{datasets}
