% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mtcars_w2lbls.R
\docType{data}
\name{mtcars_w2lbls}
\alias{mtcars_w2lbls}
\title{This is the mtcars data frame with labels added for columns mpg and cyl.}
\format{
A data frame with 32 observations and 11 columns:
\describe{
\item{ mpg }{Miles/(US) gallon}
\item{ cyl }{Number of cylinders}
\item{ disp }{Displacement (cu.in.)}
\item{ hp }{Gross horsepower}
\item{ drat }{Rear axle ratio}
\item{ wt }{Weight (1000 lbs)}
\item{ qsec }{1/4 mile time}
\item{ vs }{Engine (0 = V-shaped, 1 = straight)}
\item{ am }{Transmission (0 = automatic, 1 = manual)}
\item{ gear }{Number of forward gears}
\item{ carb }{Number of carburetors}
}
}
\source{
Generated from repfun package.
}
\usage{
mtcars_w2lbls
}
\description{
mtcars_w2lbls
}
\details{
Updated version of the mtcars data frame that includes labels for 2 variables.
}
\examples{
data("mtcars_w2lbls")
}
\references{
None
}
\keyword{datasets}
