% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rem_pg_nums.R
\name{rem_pg_nums}
\alias{rem_pg_nums}
\title{Remove floating page numbers from RTF files.}
\usage{
rem_pg_nums(infile)
}
\arguments{
\item{infile}{A path to the RTF file as a string.}
}
\value{
No return value, the file is modified.
}
\description{
Pass in an RTF file to have floating page numbers removed and result saved to the same file.
}
\examples{
\dontrun{
library(repfun)
repfun::rem_pg_nums('/path/to/an/rtf/file.rtf')
}
}
\author{
Yongwei Wang, \email{yongwei.x.wang@viivhealthcare.com} \cr
Chris Rook, \email{christopher.x.rook@gsk.com}
}
