% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ru_contents.R
\name{ru_contents}
\alias{ru_contents}
\title{Generate proc contents of data frame similar to that generated by SAS}
\usage{
ru_contents(dsetin)
}
\arguments{
\item{dsetin}{Incoming data frame to have proc contents generated.}
}
\value{
'NULL' because the proc contents output will be displayed.
}
\description{
Pass in a data frame and the contents will be displayed.
}
\examples{
library(repfun)
repfun::ru_contents(mtcars)

}
\author{
Yongwei Wang, \email{yongwei.x.wang@viivhealthcare.com} \cr
Chris Rook, \email{cr883296@gmail.com}
}
