% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ru_datacompare.R
\name{ru_datacompare}
\alias{ru_datacompare}
\title{Compare 2 data frames and report differences.}
\usage{
ru_datacompare(dsetinbase, dsetincomp, idvars, maxprint = 50)
}
\arguments{
\item{dsetinbase}{First data set.}

\item{dsetincomp}{Second data set.}

\item{idvars}{Match on these values prior to comparing records.}

\item{maxprint}{Maximum number of differences per variable to display.}
}
\value{
An output similar to proc compare will be displayed.
}
\description{
Pass in a base and compare data frame to find out if they are equal similar to proc compare in SAS.
}
\examples{
library(repfun)
repfun::ru_datacompare(iris,iris,idvars='Species')

}
\author{
Yongwei Wang, \email{yongwei.x.wang@viivhealthcare.com} \cr
Chris Rook, \email{cr883296@gmail.com}
}
