% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ru_load_library.R
\name{ru_load_library}
\alias{ru_load_library}
\title{Load a list of libraries}
\usage{
ru_load_library(pkgs)
}
\arguments{
\item{pkgs}{A list of packages to check if installed and then load.}
}
\value{
'NULL' because packages are loaded.
}
\description{
Given a list of packages, check if installed and generate message, otherwise load package.
}
\examples{
library(repfun)
repfun::ru_load_library(c("dplyr", "haven", "magrittr", "r2rtf"))

}
\author{
Yongwei Wang, \email{yongwei.x.wang@viivhealthcare.com} \cr
Chris Rook, \email{cr883296@gmail.com}
}
