% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ru_setdata.R
\name{ru_setdata}
\alias{ru_setdata}
\title{Append data sets even when variables do not match}
\usage{
ru_setdata(..., keeprownames = TRUE)
}
\arguments{
\item{...}{A collection of data frames.}

\item{keeprownames}{Convert row names on data frame to columns and keep the column.}
}
\value{
The incoming data frames combined.
}
\description{
Pass in a collection of data frames separated by commas and they will be appended.
}
\examples{
library(repfun)
repfun::ru_setdata(head(mtcars,5),head(airquality,5))

}
\author{
Yongwei Wang, \email{yongwei.x.wang@viivhealthcare.com} \cr
Chris Rook, \email{christopher.x.rook@gsk.com}
}
