% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setpath.R
\name{setpath}
\alias{setpath}
\title{Set the current working directory.}
\usage{
setpath(p)
}
\arguments{
\item{p}{A path as a string.}
}
\value{
No return value, the current working directory is set.
}
\description{
Set the current working directory so that relative paths work as expected.
}
\examples{
\dontrun{
library(repfun)
repfun::setpath()
}
}
\author{
Yongwei Wang, \email{yongwei.x.wang@viivhealthcare.com} \cr
Chris Rook, \email{christopher.x.rook@gsk.com}
}
