% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repfun-package.R
\docType{package}
\name{repfun-package}
\alias{repfun}
\alias{repfun-package}
\title{repfun: Create Tables, Listings and Figuress using Functions Styled after SAS™ Macros}
\description{
Mimic the style of traditional reporting macros for clinical trials. The purpose is to generate tables, listings and figures that support clinical research. This package is well suited for firms or individuals who wish to incorporate R without changing their ways of working as it follows a traditional clinical research workflow. Invoke functions (instead of macros) to summarize data and produce formatted reports. This package differs from others in that it includes tools (wrappers) for both analyzing and reporting data.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gsk-biostatistics.github.io/repfun/}
  \item \url{https://github.com/GSK-Biostatistics/repfun}
  \item Report bugs at \url{https://github.com/GSK-Biostatistics/repfun/issues}
}

}
\author{
\strong{Maintainer}: Chris Rook \email{cr883296@gmail.com}

Authors:
\itemize{
  \item Yongwei Wang \email{yongwei.x.wang@viivhealthcare.com}
}

Other contributors:
\itemize{
  \item GlaxoSmithKline Research & Development Limited [copyright holder, funder]
}

}
\keyword{internal}
