% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repfun-package.R
\docType{package}
\name{repfun-package}
\alias{repfun}
\alias{repfun-package}
\title{repfun: Create TLFs using R functions styled after SAS macros}
\description{
A clinical reporting toolkit of R functions that are written to mimic the style of SAS reporting macros. The purpose is to generate TLFs (tables, listings and figures) to support clinical trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://gsk-biostatistics.github.io/repfun}
}

}
\author{
\strong{Maintainer}: Yongwei Wang \email{yongwei.x.wang@viivhealthcare.com}

Authors:
\itemize{
  \item Chris Rook \email{christopher.x.rook@gsk.com}
}

}
\keyword{internal}
