% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repfun-package.R
\docType{package}
\name{repfun-package}
\alias{repfun}
\alias{repfun-package}
\title{repfun: Create TLFs using R Functions Styled after SAS Macros}
\description{
A clinical reporting toolkit of R functions that are written to mimic the style of SAS reporting macros. The purpose is to generate TLFs (tables, listings and figures) to support clinical trials.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://GSK-Biostatistics.github.io/repfun/}
  \item \url{https://github.com/GSK-Biostatistics/repfun}
}

}
\author{
\strong{Maintainer}: Chris Rook \email{cr883296@gmail.com}

Authors:
\itemize{
  \item Yongwei Wang \email{yongwei.x.wang@viivhealthcare.com}
}

Other contributors:
\itemize{
  \item GlaxoSmithKline Research & Development Limited [copyright holder, funder]
}

}
\keyword{internal}
